day02/requirements.txt
openai==2.18.0
python-dotenv==1.2.1
